requirements.txt
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
dataclasses-json==0.2.2
docutils==0.14
idna==2.8
marshmallow==3.0.0rc5
pkg-resources==0.0.0
pkginfo==1.5.0.1
Pygments==2.3.1
python-magic==0.4.15
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
tqdm==4.31.1
twine==1.13.0
urllib3==1.24.2
webencodings==0.5.1
